@@ -35,11 +35,17 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   
-  # Dependencias para la app MarketMove (a añadir cuando se integre Supabase)
-  # supabase_flutter: ^2.0.0
-  # provider: ^6.0.0
-  # go_router: ^13.0.0
-  # intl: ^0.18.0
+  # Supabase - Backend as a Service
+  supabase_flutter: ^2.8.0
+  
+  # Gestión de estado
+  provider: ^6.1.2
+  
+  # Navegación
+  go_router: ^14.6.0
+  
+  # Utilidades
+  intl: ^0.19.0
 
 dev_dependencies:
   flutter_test:
