@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'dart:math' as math;
 import '../../../shared/theme/app_theme.dart';
 
-/// Pantalla de inicio de sesión
 class LoginScreen extends StatefulWidget {
   const LoginScreen({super.key});
 
@@ -9,36 +9,87 @@ class LoginScreen extends StatefulWidget {
   State<LoginScreen> createState() => _LoginScreenState();
 }
 
-class _LoginScreenState extends State<LoginScreen> {
+class _LoginScreenState extends State<LoginScreen>
+    with TickerProviderStateMixin {
   final _formKey = GlobalKey<FormState>();
   final _emailController = TextEditingController();
   final _passwordController = TextEditingController();
   bool _obscurePassword = true;
   bool _isLoading = false;
 
+  late AnimationController _logoController;
+  late AnimationController _formController;
+  late AnimationController _bgController;
+  late Animation<double> _logoScale;
+  late Animation<double> _formSlide;
+  late Animation<double> _formFade;
+
+  @override
+  void initState() {
+    super.initState();
+    _setupAnimations();
+  }
+
+  void _setupAnimations() {
+    _logoController = AnimationController(
+      duration: const Duration(milliseconds: 1200),
+      vsync: this,
+    );
+
+    _formController = AnimationController(
+      duration: const Duration(milliseconds: 800),
+      vsync: this,
+    );
+
+    _bgController = AnimationController(
+      duration: const Duration(seconds: 15),
+      vsync: this,
+    );
+    _bgController.repeat();
+
+    _logoScale = Tween<double>(begin: 0.0, end: 1.0).animate(
+      CurvedAnimation(parent: _logoController, curve: Curves.elasticOut),
+    );
+
+    _formSlide = Tween<double>(begin: 50.0, end: 0.0).animate(
+      CurvedAnimation(parent: _formController, curve: Curves.easeOutCubic),
+    );
+
+    _formFade = Tween<double>(
+      begin: 0.0,
+      end: 1.0,
+    ).animate(CurvedAnimation(parent: _formController, curve: Curves.easeOut));
+
+    _logoController.forward().then((value) {
+      _formController.forward();
+    });
+  }
+
   @override
   void dispose() {
     _emailController.dispose();
     _passwordController.dispose();
+    _logoController.dispose();
+    _formController.dispose();
+    _bgController.dispose();
     super.dispose();
   }
 
-  void _togglePasswordVisibility() {
+  void _togglePassword() {
     setState(() {
       _obscurePassword = !_obscurePassword;
     });
   }
 
-  Future<void> _handleLogin() async {
+  Future<void> _login() async {
     if (_formKey.currentState!.validate()) {
-      setState(() => _isLoading = true);
-
-      // Simular un delay de autenticación (MVP - sin backend real)
+      setState(() {
+        _isLoading = true;
+      });
       await Future.delayed(const Duration(seconds: 1));
-
-      setState(() => _isLoading = false);
-
-      // Navegar a la pantalla principal
+      setState(() {
+        _isLoading = false;
+      });
       if (mounted) {
         Navigator.pushReplacementNamed(context, '/home');
       }
@@ -48,175 +99,173 @@ class _LoginScreenState extends State<LoginScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: SafeArea(
-        child: SingleChildScrollView(
-          padding: const EdgeInsets.all(24.0),
-          child: Form(
-            key: _formKey,
+      body: Stack(
+        children: [
+          _buildBackground(),
+          SafeArea(
+            child: SingleChildScrollView(
+              padding: const EdgeInsets.all(24),
+              child: Form(
+                key: _formKey,
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.stretch,
+                  children: [
+                    const SizedBox(height: 40),
+                    _buildLogo(),
+                    const SizedBox(height: 48),
+                    _buildForm(),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildBackground() {
+    return AnimatedBuilder(
+      animation: _bgController,
+      builder: (context, child) {
+        return CustomPaint(
+          painter: _BgPainter(_bgController.value, AppTheme.primaryColor),
+          size: Size.infinite,
+        );
+      },
+    );
+  }
+
+  Widget _buildLogo() {
+    return AnimatedBuilder(
+      animation: _logoController,
+      builder: (context, child) {
+        return Transform.scale(
+          scale: _logoScale.value,
+          child: Column(
+            children: [
+              Container(
+                width: 120,
+                height: 120,
+                decoration: BoxDecoration(
+                  gradient: LinearGradient(
+                    colors: [
+                      AppTheme.primaryColor,
+                      AppTheme.primaryColor.withValues(alpha: 0.7),
+                    ],
+                    begin: Alignment.topLeft,
+                    end: Alignment.bottomRight,
+                  ),
+                  shape: BoxShape.circle,
+                  boxShadow: [
+                    BoxShadow(
+                      color: AppTheme.primaryColor.withValues(alpha: 0.4),
+                      blurRadius: 20,
+                      offset: const Offset(0, 10),
+                    ),
+                  ],
+                ),
+                child: const Icon(
+                  Icons.store_rounded,
+                  size: 60,
+                  color: Colors.white,
+                ),
+              ),
+              const SizedBox(height: 24),
+              Text(
+                'MarketMove',
+                style: Theme.of(context).textTheme.headlineMedium?.copyWith(
+                  fontWeight: FontWeight.bold,
+                  color: AppTheme.primaryColor,
+                ),
+              ),
+              const SizedBox(height: 8),
+              Text(
+                'Gestion inteligente para tu comercio',
+                style: Theme.of(
+                  context,
+                ).textTheme.bodyMedium?.copyWith(color: AppTheme.textSecondary),
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+
+  Widget _buildForm() {
+    return AnimatedBuilder(
+      animation: _formController,
+      builder: (context, child) {
+        return Transform.translate(
+          offset: Offset(0, _formSlide.value),
+          child: Opacity(
+            opacity: _formFade.value,
             child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
               children: [
-                const SizedBox(height: 60),
-
-                // Logo y título
-                _buildHeader(),
-
-                const SizedBox(height: 48),
-
-                // Campo de email
                 _buildEmailField(),
-
                 const SizedBox(height: 16),
-
-                // Campo de contraseña
                 _buildPasswordField(),
-
                 const SizedBox(height: 8),
-
-                // Olvidé mi contraseña
-                _buildForgotPassword(),
-
+                Align(
+                  alignment: Alignment.centerRight,
+                  child: TextButton(
+                    onPressed: () {},
+                    child: const Text('Olvidaste tu password?'),
+                  ),
+                ),
                 const SizedBox(height: 24),
-
-                // Botón de login
                 _buildLoginButton(),
-
-                const SizedBox(height: 24),
-
-                // Separador
+                const SizedBox(height: 32),
                 _buildDivider(),
-
                 const SizedBox(height: 24),
-
-                // Link a registro
+                _buildSocialButtons(),
+                const SizedBox(height: 32),
                 _buildRegisterLink(),
               ],
             ),
           ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildHeader() {
-    return Column(
-      children: [
-        // Icono/Logo
-        Container(
-          width: 100,
-          height: 100,
-          decoration: BoxDecoration(
-            color: AppTheme.primaryColor.withOpacity(0.1),
-            shape: BoxShape.circle,
-          ),
-          child: const Icon(
-            Icons.store,
-            size: 50,
-            color: AppTheme.primaryColor,
-          ),
-        ),
-        const SizedBox(height: 24),
-        Text(
-          'MarketMove',
-          style: Theme.of(context).textTheme.headlineMedium?.copyWith(
-            color: AppTheme.primaryColor,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        const SizedBox(height: 8),
-        Text(
-          'Gestión para tu comercio',
-          style: Theme.of(
-            context,
-          ).textTheme.bodyMedium?.copyWith(color: AppTheme.textSecondary),
-        ),
-      ],
+        );
+      },
     );
   }
 
   Widget _buildEmailField() {
-    return TextFormField(
+    return _AnimatedField(
       controller: _emailController,
+      label: 'Correo electronico',
+      hint: 'ejemplo@correo.com',
+      icon: Icons.email_outlined,
       keyboardType: TextInputType.emailAddress,
-      textInputAction: TextInputAction.next,
-      decoration: const InputDecoration(
-        labelText: 'Correo electrónico',
-        hintText: 'ejemplo@correo.com',
-        prefixIcon: Icon(Icons.email_outlined),
-      ),
-      validator: (value) {
-        if (value == null || value.isEmpty) {
-          return 'Por favor ingresa tu correo';
-        }
-        if (!value.contains('@')) {
-          return 'Ingresa un correo válido';
-        }
+      validator: (v) {
+        if (v == null || v.isEmpty) return 'Ingresa tu correo';
+        if (!v.contains('@')) return 'Correo invalido';
         return null;
       },
     );
   }
 
   Widget _buildPasswordField() {
-    return TextFormField(
+    return _AnimatedField(
       controller: _passwordController,
-      obscureText: _obscurePassword,
-      textInputAction: TextInputAction.done,
-      decoration: InputDecoration(
-        labelText: 'Contraseña',
-        hintText: '••••••••',
-        prefixIcon: const Icon(Icons.lock_outlined),
-        suffixIcon: IconButton(
-          icon: Icon(
-            _obscurePassword ? Icons.visibility_off : Icons.visibility,
-          ),
-          onPressed: _togglePasswordVisibility,
-        ),
+      label: 'Password',
+      hint: '********',
+      icon: Icons.lock_outlined,
+      obscure: _obscurePassword,
+      suffixIcon: IconButton(
+        icon: Icon(_obscurePassword ? Icons.visibility_off : Icons.visibility),
+        onPressed: _togglePassword,
       ),
-      validator: (value) {
-        if (value == null || value.isEmpty) {
-          return 'Por favor ingresa tu contraseña';
-        }
-        if (value.length < 6) {
-          return 'La contraseña debe tener al menos 6 caracteres';
-        }
+      validator: (v) {
+        if (v == null || v.isEmpty) return 'Ingresa tu password';
+        if (v.length < 6) return 'Minimo 6 caracteres';
         return null;
       },
-      onFieldSubmitted: (_) => _handleLogin(),
-    );
-  }
-
-  Widget _buildForgotPassword() {
-    return Align(
-      alignment: Alignment.centerRight,
-      child: TextButton(
-        onPressed: () {
-          // TODO: Implementar recuperación de contraseña
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(content: Text('Función en desarrollo')),
-          );
-        },
-        child: const Text('¿Olvidaste tu contraseña?'),
-      ),
     );
   }
 
   Widget _buildLoginButton() {
-    return SizedBox(
-      height: 50,
-      child: ElevatedButton(
-        onPressed: _isLoading ? null : _handleLogin,
-        child: _isLoading
-            ? const SizedBox(
-                width: 24,
-                height: 24,
-                child: CircularProgressIndicator(
-                  strokeWidth: 2,
-                  valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
-                ),
-              )
-            : const Text('Iniciar Sesión'),
-      ),
-    );
+    return _LoginButton(isLoading: _isLoading, onPressed: _login);
   }
 
   Widget _buildDivider() {
@@ -225,9 +274,23 @@ class _LoginScreenState extends State<LoginScreen> {
         Expanded(child: Divider(color: Colors.grey.shade300)),
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: Text('o', style: TextStyle(color: Colors.grey.shade500)),
+          child: Text(
+            'o continua con',
+            style: TextStyle(color: Colors.grey.shade500),
+          ),
         ),
         Expanded(child: Divider(color: Colors.grey.shade300)),
+      ],
+    );
+  }
+
+  Widget _buildSocialButtons() {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        _SocialBtn(icon: Icons.g_mobiledata, color: Colors.red),
+        const SizedBox(width: 16),
+        _SocialBtn(icon: Icons.apple, color: Colors.black),
       ],
     );
   }
@@ -236,14 +299,277 @@ class _LoginScreenState extends State<LoginScreen> {
     return Row(
       mainAxisAlignment: MainAxisAlignment.center,
       children: [
-        const Text('¿No tienes cuenta?'),
+        const Text('No tienes cuenta?'),
         TextButton(
           onPressed: () {
             Navigator.pushNamed(context, '/register');
           },
-          child: const Text('Regístrate'),
+          child: const Text(
+            'Registrate',
+            style: TextStyle(fontWeight: FontWeight.bold),
+          ),
         ),
       ],
     );
+  }
+}
+
+class _AnimatedField extends StatefulWidget {
+  final TextEditingController controller;
+  final String label;
+  final String hint;
+  final IconData icon;
+  final Widget? suffixIcon;
+  final bool obscure;
+  final TextInputType? keyboardType;
+  final String? Function(String?)? validator;
+
+  const _AnimatedField({
+    required this.controller,
+    required this.label,
+    required this.hint,
+    required this.icon,
+    this.suffixIcon,
+    this.obscure = false,
+    this.keyboardType,
+    this.validator,
+  });
+
+  @override
+  State<_AnimatedField> createState() => _AnimatedFieldState();
+}
+
+class _AnimatedFieldState extends State<_AnimatedField>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _ctrl;
+  late Animation<double> _scale;
+  bool _focused = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _ctrl = AnimationController(
+      duration: const Duration(milliseconds: 400),
+      vsync: this,
+    );
+    _scale = Tween<double>(
+      begin: 0.9,
+      end: 1.0,
+    ).animate(CurvedAnimation(parent: _ctrl, curve: Curves.elasticOut));
+    _ctrl.forward();
+  }
+
+  @override
+  void dispose() {
+    _ctrl.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return ScaleTransition(
+      scale: _scale,
+      child: AnimatedContainer(
+        duration: const Duration(milliseconds: 200),
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(12),
+          boxShadow: _focused
+              ? [
+                  BoxShadow(
+                    color: AppTheme.primaryColor.withValues(alpha: 0.2),
+                    blurRadius: 10,
+                    offset: const Offset(0, 4),
+                  ),
+                ]
+              : [],
+        ),
+        child: Focus(
+          onFocusChange: (f) {
+            setState(() {
+              _focused = f;
+            });
+          },
+          child: TextFormField(
+            controller: widget.controller,
+            obscureText: widget.obscure,
+            keyboardType: widget.keyboardType,
+            decoration: InputDecoration(
+              labelText: widget.label,
+              hintText: widget.hint,
+              prefixIcon: Icon(widget.icon),
+              suffixIcon: widget.suffixIcon,
+              filled: true,
+              fillColor: Colors.white,
+            ),
+            validator: widget.validator,
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _LoginButton extends StatefulWidget {
+  final bool isLoading;
+  final VoidCallback onPressed;
+
+  const _LoginButton({required this.isLoading, required this.onPressed});
+
+  @override
+  State<_LoginButton> createState() => _LoginButtonState();
+}
+
+class _LoginButtonState extends State<_LoginButton>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _ctrl;
+  late Animation<double> _scale;
+
+  @override
+  void initState() {
+    super.initState();
+    _ctrl = AnimationController(
+      duration: const Duration(milliseconds: 100),
+      vsync: this,
+    );
+    _scale = Tween<double>(
+      begin: 1.0,
+      end: 0.95,
+    ).animate(CurvedAnimation(parent: _ctrl, curve: Curves.easeInOut));
+  }
+
+  @override
+  void dispose() {
+    _ctrl.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTapDown: (details) {
+        _ctrl.forward();
+      },
+      onTapUp: (details) {
+        _ctrl.reverse();
+        if (!widget.isLoading) {
+          widget.onPressed();
+        }
+      },
+      onTapCancel: () {
+        _ctrl.reverse();
+      },
+      child: ScaleTransition(
+        scale: _scale,
+        child: AnimatedContainer(
+          duration: const Duration(milliseconds: 300),
+          height: 56,
+          width: widget.isLoading ? 56 : double.infinity,
+          decoration: BoxDecoration(
+            gradient: LinearGradient(
+              colors: [
+                AppTheme.primaryColor,
+                AppTheme.primaryColor.withValues(alpha: 0.8),
+              ],
+            ),
+            borderRadius: BorderRadius.circular(widget.isLoading ? 28 : 12),
+            boxShadow: [
+              BoxShadow(
+                color: AppTheme.primaryColor.withValues(alpha: 0.4),
+                blurRadius: 15,
+                offset: const Offset(0, 8),
+              ),
+            ],
+          ),
+          child: Center(
+            child: widget.isLoading
+                ? const SizedBox(
+                    width: 24,
+                    height: 24,
+                    child: CircularProgressIndicator(
+                      strokeWidth: 2,
+                      valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+                    ),
+                  )
+                : const Text(
+                    'Iniciar Sesion',
+                    style: TextStyle(
+                      color: Colors.white,
+                      fontSize: 16,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _SocialBtn extends StatelessWidget {
+  final IconData icon;
+  final Color color;
+
+  const _SocialBtn({required this.icon, required this.color});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: 56,
+      height: 56,
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(12),
+        border: Border.all(color: Colors.grey.shade200),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.grey.shade200,
+            blurRadius: 8,
+            offset: const Offset(0, 4),
+          ),
+        ],
+      ),
+      child: Icon(icon, color: color, size: 32),
+    );
+  }
+}
+
+class _BgPainter extends CustomPainter {
+  final double anim;
+  final Color color;
+
+  _BgPainter(this.anim, this.color);
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    final paint = Paint()
+      ..style = PaintingStyle.fill
+      ..shader = LinearGradient(
+        begin: Alignment.topLeft,
+        end: Alignment.bottomRight,
+        colors: [
+          color.withValues(alpha: 0.05),
+          Colors.white,
+          color.withValues(alpha: 0.02),
+        ],
+      ).createShader(Rect.fromLTWH(0, 0, size.width, size.height));
+
+    canvas.drawRect(Rect.fromLTWH(0, 0, size.width, size.height), paint);
+
+    final cp = Paint()
+      ..style = PaintingStyle.fill
+      ..color = color.withValues(alpha: 0.05);
+
+    final x1 = size.width * 0.8 + math.sin(anim * 2 * math.pi) * 20;
+    final y1 = size.height * 0.2 + math.cos(anim * 2 * math.pi) * 20;
+    canvas.drawCircle(Offset(x1, y1), 100, cp);
+
+    final x2 = size.width * 0.2 + math.cos(anim * 2 * math.pi) * 30;
+    final y2 = size.height * 0.8 + math.sin(anim * 2 * math.pi) * 30;
+    canvas.drawCircle(Offset(x2, y2), 80, cp);
+  }
+
+  @override
+  bool shouldRepaint(_BgPainter old) {
+    return old.anim != anim;
   }
 }
