@@ -5,7 +5,6 @@ import '../../gastos/screens/gastos_screen.dart';
 import '../../productos/screens/productos_screen.dart';
 import '../../resumen/screens/resumen_screen.dart';
 
-/// Pantalla principal con navegación inferior
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
 
@@ -13,10 +12,12 @@ class HomeScreen extends StatefulWidget {
   State<HomeScreen> createState() => _HomeScreenState();
 }
 
-class _HomeScreenState extends State<HomeScreen> {
+class _HomeScreenState extends State<HomeScreen> with TickerProviderStateMixin {
   int _currentIndex = 0;
+  late PageController _pageController;
+  late AnimationController _fabController;
+  late Animation<double> _fabScale;
 
-  // Lista de pantallas para cada tab
   final List<Widget> _screens = const [
     ResumenScreen(),
     VentasScreen(),
@@ -24,55 +25,239 @@ class _HomeScreenState extends State<HomeScreen> {
     ProductosScreen(),
   ];
 
-  // Títulos para el AppBar
   final List<String> _titles = ['Resumen', 'Ventas', 'Gastos', 'Productos'];
+
+  final List<IconData> _icons = [
+    Icons.dashboard_outlined,
+    Icons.point_of_sale_outlined,
+    Icons.receipt_long_outlined,
+    Icons.inventory_2_outlined,
+  ];
+
+  final List<IconData> _activeIcons = [
+    Icons.dashboard,
+    Icons.point_of_sale,
+    Icons.receipt_long,
+    Icons.inventory_2,
+  ];
+
+  @override
+  void initState() {
+    super.initState();
+    _pageController = PageController();
+    _fabController = AnimationController(
+      duration: const Duration(milliseconds: 300),
+      vsync: this,
+    );
+    _fabScale = Tween<double>(begin: 0.0, end: 1.0).animate(
+      CurvedAnimation(parent: _fabController, curve: Curves.elasticOut),
+    );
+    _fabController.forward();
+  }
+
+  @override
+  void dispose() {
+    _pageController.dispose();
+    _fabController.dispose();
+    super.dispose();
+  }
 
   void _onTabTapped(int index) {
     setState(() {
       _currentIndex = index;
     });
+    _pageController.animateToPage(
+      index,
+      duration: const Duration(milliseconds: 300),
+      curve: Curves.easeInOut,
+    );
+    _fabController.forward(from: 0);
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text(_titles[_currentIndex]),
-        actions: [
-          // Botón de perfil/configuración
-          IconButton(
-            icon: const Icon(Icons.account_circle),
-            onPressed: () => _showUserMenu(context),
+      appBar: _buildAnimatedAppBar(),
+      body: PageView(
+        controller: _pageController,
+        onPageChanged: (index) {
+          setState(() {
+            _currentIndex = index;
+          });
+          _fabController.forward(from: 0);
+        },
+        children: _screens,
+      ),
+      bottomNavigationBar: _buildAnimatedBottomBar(),
+      floatingActionButton: _buildAnimatedFab(),
+    );
+  }
+
+  PreferredSizeWidget _buildAnimatedAppBar() {
+    return AppBar(
+      title: AnimatedSwitcher(
+        duration: const Duration(milliseconds: 300),
+        transitionBuilder: (child, animation) {
+          return FadeTransition(
+            opacity: animation,
+            child: SlideTransition(
+              position: Tween<Offset>(
+                begin: const Offset(0, 0.5),
+                end: Offset.zero,
+              ).animate(animation),
+              child: child,
+            ),
+          );
+        },
+        child: Text(_titles[_currentIndex], key: ValueKey<int>(_currentIndex)),
+      ),
+      actions: [
+        ScaleTransition(
+          scale: _fabScale,
+          child: IconButton(
+            icon: const Icon(Icons.notifications_outlined),
+            onPressed: () {
+              _showNotifications(context);
+            },
+          ),
+        ),
+        IconButton(
+          icon: const Icon(Icons.account_circle),
+          onPressed: () {
+            _showUserMenu(context);
+          },
+        ),
+      ],
+    );
+  }
+
+  Widget _buildAnimatedBottomBar() {
+    return Container(
+      decoration: BoxDecoration(
+        color: Colors.white,
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withValues(alpha: 0.1),
+            blurRadius: 20,
+            offset: const Offset(0, -5),
           ),
         ],
       ),
-      body: IndexedStack(index: _currentIndex, children: _screens),
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: _currentIndex,
-        onTap: _onTabTapped,
-        items: const [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.dashboard_outlined),
-            activeIcon: Icon(Icons.dashboard),
-            label: 'Resumen',
+      child: SafeArea(
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 8),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceAround,
+            children: List.generate(4, (index) {
+              return _AnimatedNavItem(
+                icon: _icons[index],
+                activeIcon: _activeIcons[index],
+                label: _titles[index],
+                isSelected: _currentIndex == index,
+                onTap: () {
+                  _onTabTapped(index);
+                },
+              );
+            }),
           ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.point_of_sale_outlined),
-            activeIcon: Icon(Icons.point_of_sale),
-            label: 'Ventas',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.receipt_long_outlined),
-            activeIcon: Icon(Icons.receipt_long),
-            label: 'Gastos',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.inventory_2_outlined),
-            activeIcon: Icon(Icons.inventory_2),
-            label: 'Productos',
-          ),
-        ],
+        ),
       ),
+    );
+  }
+
+  Widget? _buildAnimatedFab() {
+    if (_currentIndex == 0) return null;
+
+    return ScaleTransition(
+      scale: _fabScale,
+      child: FloatingActionButton.extended(
+        onPressed: () {
+          _showAddDialog();
+        },
+        icon: const Icon(Icons.add),
+        label: Text(_getFabLabel()),
+        backgroundColor: _getFabColor(),
+      ),
+    );
+  }
+
+  String _getFabLabel() {
+    switch (_currentIndex) {
+      case 1:
+        return 'Nueva Venta';
+      case 2:
+        return 'Nuevo Gasto';
+      case 3:
+        return 'Nuevo Producto';
+      default:
+        return 'Nuevo';
+    }
+  }
+
+  Color _getFabColor() {
+    switch (_currentIndex) {
+      case 1:
+        return AppTheme.successColor;
+      case 2:
+        return AppTheme.errorColor;
+      case 3:
+        return AppTheme.primaryColor;
+      default:
+        return AppTheme.primaryColor;
+    }
+  }
+
+  void _showAddDialog() {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text('Agregar ${_getFabLabel()}'),
+        behavior: SnackBarBehavior.floating,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+      ),
+    );
+  }
+
+  void _showNotifications(BuildContext context) {
+    showModalBottomSheet(
+      context: context,
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
+      ),
+      builder: (context) {
+        return Container(
+          padding: const EdgeInsets.all(20),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Container(
+                width: 40,
+                height: 4,
+                margin: const EdgeInsets.only(bottom: 20),
+                decoration: BoxDecoration(
+                  color: Colors.grey.shade300,
+                  borderRadius: BorderRadius.circular(2),
+                ),
+              ),
+              const Icon(
+                Icons.notifications_none,
+                size: 48,
+                color: Colors.grey,
+              ),
+              const SizedBox(height: 16),
+              const Text(
+                'Sin notificaciones',
+                style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+              ),
+              const SizedBox(height: 8),
+              Text(
+                'Las notificaciones apareceran aqui',
+                style: TextStyle(color: Colors.grey.shade600),
+              ),
+              const SizedBox(height: 20),
+            ],
+          ),
+        );
+      },
     );
   }
 
@@ -173,9 +358,115 @@ class _HomeScreenState extends State<HomeScreen> {
               Navigator.pop(context);
               Navigator.pushReplacementNamed(context, '/login');
             },
-            child: const Text('Cerrar sesión'),
+            child: const Text('Cerrar sesion'),
           ),
         ],
+      ),
+    );
+  }
+}
+
+class _AnimatedNavItem extends StatefulWidget {
+  final IconData icon;
+  final IconData activeIcon;
+  final String label;
+  final bool isSelected;
+  final VoidCallback onTap;
+
+  const _AnimatedNavItem({
+    required this.icon,
+    required this.activeIcon,
+    required this.label,
+    required this.isSelected,
+    required this.onTap,
+  });
+
+  @override
+  State<_AnimatedNavItem> createState() => _AnimatedNavItemState();
+}
+
+class _AnimatedNavItemState extends State<_AnimatedNavItem>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _controller;
+  late Animation<double> _scaleAnimation;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = AnimationController(
+      duration: const Duration(milliseconds: 200),
+      vsync: this,
+    );
+    _scaleAnimation = Tween<double>(
+      begin: 1.0,
+      end: 0.9,
+    ).animate(CurvedAnimation(parent: _controller, curve: Curves.easeInOut));
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTapDown: (d) {
+        _controller.forward();
+      },
+      onTapUp: (d) {
+        _controller.reverse();
+        widget.onTap();
+      },
+      onTapCancel: () {
+        _controller.reverse();
+      },
+      child: ScaleTransition(
+        scale: _scaleAnimation,
+        child: AnimatedContainer(
+          duration: const Duration(milliseconds: 200),
+          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+          decoration: BoxDecoration(
+            color: widget.isSelected
+                ? AppTheme.primaryColor.withValues(alpha: 0.1)
+                : Colors.transparent,
+            borderRadius: BorderRadius.circular(20),
+          ),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              AnimatedSwitcher(
+                duration: const Duration(milliseconds: 200),
+                transitionBuilder: (child, animation) {
+                  return ScaleTransition(scale: animation, child: child);
+                },
+                child: Icon(
+                  widget.isSelected ? widget.activeIcon : widget.icon,
+                  key: ValueKey<bool>(widget.isSelected),
+                  color: widget.isSelected
+                      ? AppTheme.primaryColor
+                      : Colors.grey,
+                  size: widget.isSelected ? 28 : 24,
+                ),
+              ),
+              const SizedBox(height: 4),
+              AnimatedDefaultTextStyle(
+                duration: const Duration(milliseconds: 200),
+                style: TextStyle(
+                  color: widget.isSelected
+                      ? AppTheme.primaryColor
+                      : Colors.grey,
+                  fontSize: widget.isSelected ? 12 : 11,
+                  fontWeight: widget.isSelected
+                      ? FontWeight.bold
+                      : FontWeight.normal,
+                ),
+                child: Text(widget.label),
+              ),
+            ],
+          ),
+        ),
       ),
     );
   }
